minor: move notification methods from app.tsx to notif service
